Move CopyOptimizer to new module
